feat: cleaning up product-item service
